+++ example/audio-element/main.js
@@ -10,7 +10,12 @@ document.addEventListener('DOMContentLoaded', function() {
         container: document.querySelector('#waveform'),
         waveColor: '#A8DBA8',
         progressColor: '#3B8686',
-        backend: 'MediaElement'
+        backend: 'MediaElement',
+        mediaControls: false
+    });
+
+    wavesurfer.once('ready', function() {
+        console.log('Using wavesurfer.js ' + WaveSurfer.VERSION);
     });
 
     // Load audio from URL